+++ bootcfg/http/handlers_test.go
@@ -15,30 +15,6 @@ import (
 	fake "github.com/coreos/coreos-baremetal/bootcfg/storage/testfakes"
 )
 
-func TestRequireGET(t *testing.T) {
-	next := func(w http.ResponseWriter, req *http.Request) {
-		fmt.Fprintf(w, "next")
-	}
-	h := requireGET(http.HandlerFunc(next))
-	req, _ := http.NewRequest("GET", "/", nil)
-	w := httptest.NewRecorder()
-	h.ServeHTTP(w, req)
-	assert.Equal(t, http.StatusOK, w.Code)
-	assert.Equal(t, "next", w.Body.String())
-}
-
-func TestRequireGET_WrongMethod(t *testing.T) {
-	next := func(w http.ResponseWriter, req *http.Request) {
-		fmt.Fprintf(w, "next")
-	}
-	h := requireGET(http.HandlerFunc(next))
-	req, _ := http.NewRequest("POST", "/", nil)
-	w := httptest.NewRecorder()
-	h.ServeHTTP(w, req)
-	assert.Equal(t, http.StatusMethodNotAllowed, w.Code)
-	assert.Equal(t, "only HTTP GET is supported\n", w.Body.String())
-}
-
 func TestSelectGroup(t *testing.T) {
 	store := &fake.FixedStore{
 		Groups: map[string]*storagepb.Group{fake.Group.Id: fake.Group},
