+++ bootcfg/http/server.go
@@ -48,12 +48,12 @@ func (s *Server) HTTPHandler() http.Handler {
 		return s.logRequest(NewHandler(next))
 	}
 	// bootcfg version
-	mux.Handle("/", s.logRequest(versionHandler()))
+	mux.Handle("/", s.logRequest(homeHandler()))
 	// Boot via GRUB
 	mux.Handle("/grub", chain(s.selectProfile(s.core, s.grubHandler())))
 	// Boot via iPXE
-	mux.Handle("/boot.ipxe", s.logRequest(ipxeInspect()))
-	mux.Handle("/boot.ipxe.0", s.logRequest(ipxeInspect()))
+	mux.Handle("/boot.ipxe", chain(ipxeInspect()))
+	mux.Handle("/boot.ipxe.0", chain(ipxeInspect()))
 	mux.Handle("/ipxe", chain(s.selectProfile(s.core, s.ipxeHandler())))
 	// Boot via Pixiecore
 	mux.Handle("/pixiecore/v1/boot/", chain(s.pixiecoreHandler(s.core)))
@@ -68,32 +68,32 @@ func (s *Server) HTTPHandler() http.Handler {
 
 	// Signatures
 	if s.signer != nil {
-		signerChain := func(next http.Handler) http.Handler {
-			return s.logRequest(sign.SignatureHandler(s.signer, next))
+		signerChain := func(next ContextHandler) http.Handler {
+			return s.logRequest(sign.SignatureHandler(s.signer, NewHandler(next)))
 		}
-		mux.Handle("/grub.sig", signerChain(NewHandler(s.selectProfile(s.core, s.grubHandler()))))
+		mux.Handle("/grub.sig", signerChain(s.selectProfile(s.core, s.grubHandler())))
 		mux.Handle("/boot.ipxe.sig", signerChain(ipxeInspect()))
 		mux.Handle("/boot.ipxe.0.sig", signerChain(ipxeInspect()))
-		mux.Handle("/ipxe.sig", signerChain(NewHandler(s.selectProfile(s.core, s.ipxeHandler()))))
-		mux.Handle("/pixiecore/v1/boot.sig/", signerChain(NewHandler(s.pixiecoreHandler(s.core))))
-		mux.Handle("/ignition.sig", signerChain(NewHandler(s.selectGroup(s.core, s.ignitionHandler(s.core)))))
-		mux.Handle("/cloud.sig", signerChain(NewHandler(s.selectGroup(s.core, s.cloudHandler(s.core)))))
-		mux.Handle("/generic.sig", signerChain(NewHandler(s.selectGroup(s.core, s.genericHandler(s.core)))))
-		mux.Handle("/metadata.sig", signerChain(NewHandler(s.selectGroup(s.core, s.metadataHandler()))))
+		mux.Handle("/ipxe.sig", signerChain(s.selectProfile(s.core, s.ipxeHandler())))
+		mux.Handle("/pixiecore/v1/boot.sig/", signerChain(s.pixiecoreHandler(s.core)))
+		mux.Handle("/ignition.sig", signerChain(s.selectGroup(s.core, s.ignitionHandler(s.core))))
+		mux.Handle("/cloud.sig", signerChain(s.selectGroup(s.core, s.cloudHandler(s.core))))
+		mux.Handle("/generic.sig", signerChain(s.selectGroup(s.core, s.genericHandler(s.core))))
+		mux.Handle("/metadata.sig", signerChain(s.selectGroup(s.core, s.metadataHandler())))
 	}
 	if s.armoredSigner != nil {
-		signerChain := func(next http.Handler) http.Handler {
-			return s.logRequest(sign.SignatureHandler(s.armoredSigner, next))
+		signerChain := func(next ContextHandler) http.Handler {
+			return s.logRequest(sign.SignatureHandler(s.armoredSigner, NewHandler(next)))
 		}
-		mux.Handle("/grub.asc", signerChain(NewHandler(s.selectProfile(s.core, s.grubHandler()))))
+		mux.Handle("/grub.asc", signerChain(s.selectProfile(s.core, s.grubHandler())))
 		mux.Handle("/boot.ipxe.asc", signerChain(ipxeInspect()))
 		mux.Handle("/boot.ipxe.0.asc", signerChain(ipxeInspect()))
-		mux.Handle("/ipxe.asc", signerChain(NewHandler(s.selectProfile(s.core, s.ipxeHandler()))))
-		mux.Handle("/pixiecore/v1/boot.asc/", signerChain(NewHandler(s.pixiecoreHandler(s.core))))
-		mux.Handle("/ignition.asc", signerChain(NewHandler(s.selectGroup(s.core, s.ignitionHandler(s.core)))))
-		mux.Handle("/cloud.asc", signerChain(NewHandler(s.selectGroup(s.core, s.cloudHandler(s.core)))))
-		mux.Handle("/generic.asc", signerChain(NewHandler(s.selectGroup(s.core, s.genericHandler(s.core)))))
-		mux.Handle("/metadata.asc", signerChain(NewHandler(s.selectGroup(s.core, s.metadataHandler()))))
+		mux.Handle("/ipxe.asc", signerChain(s.selectProfile(s.core, s.ipxeHandler())))
+		mux.Handle("/pixiecore/v1/boot.asc/", signerChain(s.pixiecoreHandler(s.core)))
+		mux.Handle("/ignition.asc", signerChain(s.selectGroup(s.core, s.ignitionHandler(s.core))))
+		mux.Handle("/cloud.asc", signerChain(s.selectGroup(s.core, s.cloudHandler(s.core))))
+		mux.Handle("/generic.asc", signerChain(s.selectGroup(s.core, s.genericHandler(s.core))))
+		mux.Handle("/metadata.asc", signerChain(s.selectGroup(s.core, s.metadataHandler())))
 	}
 
 	// kernel, initrd, and TLS assets
